fix the type error

+++ examples/shooter/three_modes_shooter.cc
@@ -31,7 +31,7 @@
 // #define LASER_GPIO_Port GPIOG
 
 static const int DELAY = 350;  // ms
-static const int TARGET_SPEED_FLYWHEELS = 50;
+static const float TARGET_SPEED_FLYWHEELS = 50;
 
 bsp::CAN* can = nullptr;
 control::MotorCANBase* left_flywheel_motor = nullptr;
